+++ src/main/java/com/bertramlabs/plugins/hcl4j/symbols/GenericSymbol.java
@@ -15,13 +15,13 @@
  */
 package com.bertramlabs.plugins.hcl4j.symbols;
 
-import java.util.List;
 import java.util.ArrayList;
+import java.util.List;
 
 public abstract class GenericSymbol implements Symbol {
 	private Integer line;
 	private Integer column;
-	private Integer position;
+	private Long position;
 	private Integer length;
 
 	private String name;
@@ -54,7 +54,7 @@ public abstract class GenericSymbol implements Symbol {
 		return column;
 	}
 
-	public Integer getPosition() {
+	public Long getPosition() {
 		return position;
 	}
 
@@ -86,7 +86,7 @@ public abstract class GenericSymbol implements Symbol {
 		this.name = name;
 	}
 
-	public GenericSymbol(String name,Integer line, Integer column,Integer position) {
+	public GenericSymbol(String name,Integer line, Integer column,Long position) {
 		this.name = name;
 		this.line = line;
 		this.column = column;
