+++ src/main/java/com/bertramlabs/plugins/hcl4j/symbols/HCLMap.java
@@ -15,13 +15,10 @@
  */
 package com.bertramlabs.plugins.hcl4j.symbols;
 
-import java.util.LinkedHashMap;
-import java.util.Map;
-
 public class HCLMap extends GenericSymbol {
 
 
-	public HCLMap(Integer line, Integer column,Integer position) {
+	public HCLMap(Integer line, Integer column,Long position) {
 		super("array",line,column,position);
 	}
